refactor: challenge endpoint code
