+++ crates/server/src/admin/handlers.rs
@@ -1,2 +1,3 @@
 pub mod add_client_key;
+pub mod challenge;
 pub mod fetch_did;
